+++ api/src/routes/user.js
@@ -12,6 +12,7 @@ const { canEditUser } = require('../passport')
 const db = require('../db/connection')
 const getCountriesEdited = require('../utils/getCountriesEdited')
 const refreshStatus = require('../utils/osmesaStatus.js')
+const getSumEdits = require('../utils/sum_edits')
 const { prop } = require('ramda')
 
 /**
@@ -93,7 +94,7 @@ async function get (req, res) {
   if (osmesaData.extent_uri) {
     osmesaData.extent_uri = join(APP_URL_FINAL, '/scoreboard/api/extents/', osmesaData.extent_uri)
   }
-
+  osmesaData['edit_sum'] = getSumEdits(osmesaData)
   const refreshDate = await refreshStatus('user')
 
   let countriesEdited = getCountriesEdited(osmesaData.country_list)