+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Campaigns can be sorted by name in alphabetical order
 - "Reset filters" button clears filter options on the Campaigns page
-- Edits and Changesets available in activity widget
+- Number of edits and Number of changesets now available in summary stats widget at the top of the user and dashboard pages
 
 ### Fixed
 - State retains page number and filters when navigating between the Campaigns page and individual campaign cards