+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - State retains page number and filters when navigating between the Campaigns page and individual campaign cards
+- Add uniqueness constraint to campaigns to make sure that campaigns with the same id from the same tasking manager aren't added twice
 
 ## [v1.2.0] - 2019-05-08
 ### Added